:recycle: Rename to libhal-esp8266
Resolves #31
