--- CMakeLists.txt
@@ -1,14 +1,0 @@
-cmake_minimum_required(VERSION 3.15)
-
-project(libesp8266 VERSION 0.0.1 LANGUAGES CXX)
-add_library(${PROJECT_NAME} INTERFACE)
-target_include_directories(${PROJECT_NAME} INTERFACE include)
-target_compile_features(${PROJECT_NAME} INTERFACE cxx_std_20)
-
-install(TARGETS ${PROJECT_NAME}
-        EXPORT ${PROJECT_NAME}
-        LIBRARY DESTINATION lib
-        ARCHIVE DESTINATION lib
-        RUNTIME DESTINATION bin
-        PUBLIC_HEADER DESTINATION include
-)
