+++ demos/applications/wlan_client.cpp
@@ -1,6 +1,6 @@
-#include <libesp8266/at/socket.hpp>
-#include <libesp8266/at/wlan_client.hpp>
-#include <libesp8266/util.hpp>
+#include <libhal-esp8266/at/socket.hpp>
+#include <libhal-esp8266/at/wlan_client.hpp>
+#include <libhal-esp8266/util.hpp>
 #include <libhal-util/serial.hpp>
 #include <libhal-util/steady_clock.hpp>
 #include <libhal/timeout.hpp>
