+++ include/libhal-esp8266/util.hpp
