+++ test_package/CMakeLists.txt
@@ -1,10 +1,10 @@
 cmake_minimum_required(VERSION 3.15)
 project(test_package LANGUAGES CXX)
 
-find_package(libesp8266 REQUIRED CONFIG)
+find_package(libhal-esp8266 REQUIRED CONFIG)
 
 add_executable(${PROJECT_NAME} main.cpp)
 target_include_directories(${PROJECT_NAME} PUBLIC .)
 target_compile_features(${PROJECT_NAME} PRIVATE cxx_std_20)
 set_target_properties(${PROJECT_NAME} PROPERTIES CXX_EXTENSIONS OFF)
-target_link_libraries(${PROJECT_NAME} PRIVATE libesp8266::libesp8266)
+target_link_libraries(${PROJECT_NAME} PRIVATE libhal::esp8266)
