+++ test_package/main.cpp
@@ -3,8 +3,8 @@
 
 #include <array>
 
-#include <libesp8266/at/socket.hpp>
-#include <libesp8266/at/wlan_client.hpp>
+#include <libhal-esp8266/at/socket.hpp>
+#include <libhal-esp8266/at/wlan_client.hpp>
 #include <libhal-util/serial.hpp>
 #include <libhal-util/steady_clock.hpp>
 #include <libhal/timeout.hpp>
