+++ tests/http_response.test.cpp
@@ -1,4 +1,4 @@
-#include <libesp8266/http_response.hpp>
+#include <libhal-esp8266/http_response.hpp>
 
 #include "helpers.hpp"
